ai-engine/requirements.txt
tensorflow-cpu==2.15.0  # Much lighter
fastapi
uvicorn
python-multipart
pillow
numpy